Log analytic event of child feature via `SingleAnalyticEventActionDispatcher`
