+++ shared/src/commonMain/kotlin/org/hyperskill/app/notification/click_handling/injection/NotificationClickHandlingComponentImpl.kt
@@ -2,6 +2,7 @@ package org.hyperskill.app.notification.click_handling.injection
 
 import org.hyperskill.app.core.injection.AppGraph
 import org.hyperskill.app.core.presentation.ActionDispatcherOptions
+import org.hyperskill.app.notification.click_handling.presentation.MainNotificationClickHandlingActionDispatcher
 import org.hyperskill.app.notification.click_handling.presentation.NotificationClickHandlingActionDispatcher
 import org.hyperskill.app.notification.click_handling.presentation.NotificationClickHandlingReducer
 
@@ -11,14 +12,19 @@ internal class NotificationClickHandlingComponentImpl(
     override val notificationClickHandlingReducer: NotificationClickHandlingReducer
         get() = NotificationClickHandlingReducer()
 
-    override val notificationClickHandlingActionDispatcher: NotificationClickHandlingActionDispatcher
-        get() = NotificationClickHandlingActionDispatcher(
+    private val mainNotificationClickHandlingActionDispatcher: MainNotificationClickHandlingActionDispatcher
+        get() = MainNotificationClickHandlingActionDispatcher(
             ActionDispatcherOptions(),
-            analyticInteractor = appGraph.analyticComponent.analyticInteractor,
             currentProfileStateRepository = appGraph.profileDataComponent.currentProfileStateRepository,
             nextLearningActivityStateRepository = appGraph.stateRepositoriesComponent
                 .nextLearningActivityStateRepository,
             badgesRepository = appGraph.buildBadgesDataComponent().badgesRepository,
             logger = appGraph.loggerComponent.logger
+        )
+
+    override val notificationClickHandlingActionDispatcher: NotificationClickHandlingActionDispatcher
+        get() = NotificationClickHandlingActionDispatcher(
+            mainNotificationClickHandlingActionDispatcher,
+            appGraph.analyticComponent.analyticInteractor
         )
 }
