+++ shared/src/commonMain/kotlin/org/hyperskill/app/step_toolbar/injection/StepToolbarComponentImpl.kt
@@ -3,6 +3,7 @@ package org.hyperskill.app.step_toolbar.injection
 import org.hyperskill.app.core.injection.AppGraph
 import org.hyperskill.app.core.presentation.ActionDispatcherOptions
 import org.hyperskill.app.step.domain.model.StepRoute
+import org.hyperskill.app.step_toolbar.presentation.MainStepToolbarActionDispatcher
 import org.hyperskill.app.step_toolbar.presentation.StepToolbarActionDispatcher
 import org.hyperskill.app.step_toolbar.presentation.StepToolbarReducer
 
@@ -15,9 +16,11 @@ internal class StepToolbarComponentImpl(
 
     override val stepToolbarActionDispatcher: StepToolbarActionDispatcher
         get() = StepToolbarActionDispatcher(
-            config = ActionDispatcherOptions(),
-            topicCompletedFlow = appGraph.stepCompletionFlowDataComponent.topicCompletedFlow,
-            progressesInteractor = appGraph.buildProgressesDataComponent().progressesInteractor,
-            analyticInteractor = appGraph.analyticComponent.analyticInteractor
+            MainStepToolbarActionDispatcher(
+                config = ActionDispatcherOptions(),
+                topicCompletedFlow = appGraph.stepCompletionFlowDataComponent.topicCompletedFlow,
+                progressesInteractor = appGraph.buildProgressesDataComponent().progressesInteractor
+            ),
+            appGraph.analyticComponent.analyticInteractor
         )
 }
