+++ shared/src/commonMain/kotlin/org/hyperskill/app/streak_recovery/injection/StreakRecoveryComponentImpl.kt
@@ -2,6 +2,7 @@ package org.hyperskill.app.streak_recovery.injection
 
 import org.hyperskill.app.core.injection.AppGraph
 import org.hyperskill.app.core.presentation.ActionDispatcherOptions
+import org.hyperskill.app.streak_recovery.presentation.MainStreakRecoveryActionDispatcher
 import org.hyperskill.app.streak_recovery.presentation.StreakRecoveryActionDispatcher
 import org.hyperskill.app.streak_recovery.presentation.StreakRecoveryReducer
 
@@ -11,16 +12,21 @@ internal class StreakRecoveryComponentImpl(
     override val streakRecoveryReducer: StreakRecoveryReducer
         get() = StreakRecoveryReducer(resourceProvider = appGraph.commonComponent.resourceProvider)
 
-    override val streakRecoveryActionDispatcher: StreakRecoveryActionDispatcher
-        get() = StreakRecoveryActionDispatcher(
+    private val mainStreakRecoveryActionDispatcher: MainStreakRecoveryActionDispatcher
+        get() = MainStreakRecoveryActionDispatcher(
             ActionDispatcherOptions(),
             currentProfileStateRepository = appGraph.profileDataComponent.currentProfileStateRepository,
             streaksInteractor = appGraph.buildStreaksDataComponent().streaksInteractor,
             productsInteractor = appGraph.buildProductsDataComponent().productsInteractor,
-            analyticInteractor = appGraph.analyticComponent.analyticInteractor,
             sentryInteractor = appGraph.sentryComponent.sentryInteractor,
             logger = appGraph.loggerComponent.logger,
             streakFlow = appGraph.streakFlowDataComponent.streakFlow,
             resourceProvider = appGraph.commonComponent.resourceProvider
+        )
+
+    override val streakRecoveryActionDispatcher: StreakRecoveryActionDispatcher
+        get() = StreakRecoveryActionDispatcher(
+            mainStreakRecoveryActionDispatcher,
+            appGraph.analyticComponent.analyticInteractor
         )
 }
